Upgrade to Dolittle v4
The main change is in startup where the ordering of middleware
has changed. The .UseDolittle() -method on the IApplicationBuilder
must now appear before the .UseMvc() -method.

There is also a change in how the app determines which Tenant it
is running for, and an extension is provided to set the tenant in the
development environment.

The way we use the aggregate roots has changed from the previous
versions. The aggregate root is now gotten or rehydrated and given
an action to perform. The result of this performing this action is a
result which may contain broken rules that stopped the action in the
aggregate root. This lets the command-handler react to such non-
performance if it needs to (i.e. to rollback other systems, etc)